refactor: icon 위치 변경
